# Hotfix run each request in bound thread