Fixes a problem with recently-introduced thread-local storage in the LLVM backend.
We have to make sure that all foreign calls relating to an LLVM-based term evaluation use the same OS thread, which can only be achieved [via bound threads](https://downloads.haskell.org/ghc/latest/docs/libraries/base-4.20.0.0-1f57/Control-Concurrent.html#g:8) .

This PR
* adds a flag to the generic RPC server in `kore-rpc-types` to request processing requests in bound threads
* `kore-rpc-booster` uses this flag for bound threads when an LLVM backend library is used

This _should_ protect us against problems related to using thread-local storage in the LLVM backend. Needs to be thoroughly tested before merging, because issues only materialised in proofs with substantial workload and parallel exploration.

[Branch HOTFIX-use-bound-threads-for-request-workers](https://github.com/runtimeverification/haskell-backend/compare/HOTFIX-use-bound-threads-for-request-workers) is a simple take on this, running the whole worker thread in the server as a bound thread. This version here only runs the single request in a bound thread, and also declares the LLVM calls `unsafe` because they won't read HS heap data and never call back into Haskell.